refactor: move the getLinkResolverIdentifier function to dlr service
