+++ packages/services/src/__tests__/gs1.test.ts
@@ -118,36 +118,4 @@ describe('Gs1Provider', () => {
     });
   });
 
-  describe('getLinkResolverIdentifier', () => {
-    it('should extract identifier and qualifier path from an gtin AI', () => {
-      const { identifier, qualifierPath } = gs1Provider.getLinkResolverIdentifier([
-        { ai: '01', value: '09359502000010' }
-      ]);
-
-      expect(identifier).toBe('09359502000010');
-      expect(qualifierPath).toBe('');
-    });
-
-    it('should extract identifier and qualifier path from a combined multi AIs', () => {
-      const { identifier, qualifierPath } = gs1Provider.getLinkResolverIdentifier([
-        { ai: '01', value: '09359502000010' },
-        { ai: '10', value: 'ABC123' }
-      ]);
-
-      expect(identifier).toBe('09359502000010');
-      expect(qualifierPath).toBe('/10/ABC123');
-    });
-
-    it('should throw an invalid DLR AIs error if input is empty array', () => {
-      expect(() => gs1Provider.getLinkResolverIdentifier([])).toThrow('Invalid DLR AIs. At least one DLR AI is required to resolve the identifier.');
-    });
-
-    it('should throw an invalid DLR AIs error if input 01 and 8006 are primary keys present at the same time.', () => {
-      expect(() => gs1Provider.getLinkResolverIdentifier([
-        { ai: '01', value: '09359502000010' },
-        { ai: '8006', value: 'ABC123' }
-      ])).toThrow('Invalid DLR AIs. Both 01 and 8006 are primary keys and cannot be present at the same time.');
-    });
-  });
-
 });
